# flavor-entertainers-backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Flavor Entertainers API
  description: Production-ready backend for adult entertainment performer bookings in Western Australia
  version: 1.0.0
  contact:
    name: Flavor Entertainers
    email: contact@lustandlace.com.au
  license:
    name: UNLICENSED
servers:
  - url: https://api.flavor-entertainers.example
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid request body"
        code:
          type: string
          example: "VALIDATION_FAILED"
        details:
          type: object
        requestId:
          type: string
          format: uuid
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        performer_id:
          type: string
          format: uuid
        event_date:
          type: string
          format: date
        event_time:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
        location:
          type: string
        service:
          type: string
          enum: [Topless Waitress, Striptease, XXX Show]
        rate:
          type: number
          format: decimal
        booking_fee:
          type: number
          format: decimal
        total:
          type: number
          format: decimal
        message:
          type: string
        status:
          type: string
          enum: [pending_review, awaiting_payment, confirmed, completed, cancelled]
        payment_link_url:
          type: string
          format: uri
        balance_due:
          type: number
          format: decimal
        balance_due_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
            has_more:
              type: boolean
...
paths:
...
  /bookings/request:
    post:
      tags:
        - Bookings
      summary: Create booking request
      description: Create a new booking request
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - phone
                - event_date
                - event_time
                - location
                - service
                - rate
              properties:
                performer_id:
                  type: string
                  format: uuid
                name:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  pattern: '^\+61[0-9]{9}$'
                event_date:
                  type: string
                  format: date
                event_time:
                  type: string
                  pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
                location:
                  type: string
                  maxLength: 500
                service:
                  type: string
                  enum: [Topless Waitress, Striptease, XXX Show]
                rate:
                  type: number
                  minimum: 0
                message:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: Booking request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  message:
                    type: string
        '403':
          description: Booking blocked (blacklisted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking details
      description: Get details of a specific booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bookings/{id}/approve:
    post:
      tags:
        - Bookings
      summary: Approve booking
      description: Approve a booking and generate payment link (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                performer_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Booking approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  payment_link_url:
                    type: string
                    format: uri
                  status:
                    type: string
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bookings:
    get:
      tags:
        - Bookings
      summary: List bookings
      description: List bookings with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_review, awaiting_deposit, confirmed, completed, cancelled]
        - name: performer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Booking'
...